Internationalized Wordpress Theme
Through using the Polylang plugin, the Codestyling Localization plugin,
and this post
(http://wordpress.org/support/topic/plugin-polylang-how-to-internationalize-a-theme),
I was able to translate our theme so it can be viewed in Spanish and
English.  I also edited functions.php so I could use a dynamic sidebar.
The dynamic sidebar helped me use Polylang to test the
internationalization and can be removed later if need be.