Facebook Integration Commit
In this commit, I integrated our Wordpress blog with Facebook.  For the
front page, I added a Like box which shows the Facebook feed of a fake
Facebook page I created for this project.  The main.css was modified to
make the Like Box responsive.  To make the front page look better, I
added a fourth Friends Features which shows off Our Partners to
strengthen "expanding outreach."  Due to the new feature, the .mo and
.po files needed to be modified.  I also added a Facebook like button to
appear at the bottom of every page type.